fix(UC_AddParticipant): ajout de la sécurité des méthodes du controler pour l'ajout d'un participant

+++ controller/ControllerTricount.php
@@ -321,35 +321,42 @@ class ControllerTricount extends MyController{
 
     public function add_subscriber_service(){
         $user = $this->get_user_or_redirect();
+        
+        if(isset($_GET["param1"]) && isset($_POST["userId"])){
 
-        $tricount = Tricount::get_tricount_by_id($_GET["param1"]);
-        $targetUser = User::get_user_by_id($_POST["userId"]);
+            $tricount = Tricount::get_tricount_by_id($_GET["param1"]);
+            $targetUser = User::get_user_by_id($_POST["userId"]);
 
-        if(isset($_GET["param1"]) && $_GET["param1"] !== "" && is_numeric($_GET["param1"]) && !$targetUser->is_subscribed_to_tricount($_GET["param1"])){
-            $tricount->add_subscriber($targetUser);
-        }
-        else{
-            $this->redirect();
+            if(isset($_GET["param1"]) && $_GET["param1"] !== "" && is_numeric($_GET["param1"]) && !$targetUser->is_subscribed_to_tricount($_GET["param1"])){
+                $tricount->add_subscriber($targetUser);
+            }
+            else{
+                $this->redirect();
+            }
         }
     }
 
     public function remove_subscriber_service(){
         $user = $this->get_user_or_redirect();
+        if(isset($_GET["param1"]) && isset($_POST["userId"])){
 
-        $tricount = Tricount::get_tricount_by_id($_GET["param1"]);
-        $targetUser = User::get_user_by_id($_POST["userId"]);
-        $creatorOfTricount=$user->get_creator_of_tricount($tricount);
-        if(isset($_GET["param1"]) && $_GET["param1"] !== "" && is_numeric($_GET["param1"]) && $targetUser->is_subscribed_to_tricount($_GET["param1"])){
-            if(!$targetUser->has_already_paid($tricount)&&$targetUser->id!=$creatorOfTricount->id){
-                $tricount->delete_participation($targetUser->id);
-
-            }            
-        }
-        else{
+        
+            $tricount = Tricount::get_tricount_by_id($_GET["param1"]);
+            $targetUser = User::get_user_by_id($_POST["userId"]);
+            $creatorOfTricount=$user->get_creator_of_tricount($tricount);
+            if(isset($_GET["param1"]) && $_GET["param1"] !== "" && is_numeric($_GET["param1"]) && $targetUser->is_subscribed_to_tricount($_GET["param1"])){
+                if(!$targetUser->has_already_paid($tricount)&&$targetUser->id!=$creatorOfTricount->id){
+                    $tricount->delete_participation($targetUser->id);
+                }            
+            }
+            else{
+                $this->redirect();
+            }
+        } else {
             $this->redirect();
         }
-        
     }
+
 
 
 }
